Time: 19 ms (8.48%), Space: 8.3 MB (14.11%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int mySqrt(int x) {
         if(x < 2) return x;
-        for(int i = 1; i <= x; i++){
+        for(long long i = 1; i <= x; i++){
             if(i*i == x) return i;
             if(i*i > x) return i-1;
         }
